+++ src/light.h
@@ -2,7 +2,7 @@
 #define LIGHT_H
 
 void set_color(int i);
-void set_colorf(float r, float g, float b);
+void set_colorf(double r, double g, double b);
 extern int   light_switch[3];
 void	updateLights();
 void light_init();
