+++ src/rotate.h
@@ -15,7 +15,7 @@ void mouseMotion(int x, int y, KeyboardModifier modifiers );
 void mouseButton(int button, int state, int x, int y, KeyboardModifier modifiers);
 void display();
 void spin();
-void zoom(float delta);
+void zoom(double delta);
 extern int clip_item;
 
 #endif // ROTATE_H
